+++ src/main/java/blackjack/controller/BlackjackController.java
@@ -9,8 +9,8 @@ import blackjack.domain.participant.Players;
 import blackjack.dto.CardResult;
 import blackjack.dto.ProfitResult;
 import blackjack.dto.StateResult;
-import blackjack.view.InputView;
-import blackjack.view.OutputView;
+import blackjack.view.input.InputView;
+import blackjack.view.output.OutputView;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
